Rescue failed JSON.parse so we bubble up a proper HTTP exception
Currently we’re trusting the response body to contain valid JSON. That
might not always be the case, e.g. think about an nginx error page, or
a temporary/permanent response from a CDN.

By rescuing the `JSON.parse`, we can return an empty hash and make sure
that the actual error is bubbled up to the user, not a
`JSON::ParserError`.